add scheduler updater for notification settings and upd menu

+++ src/main/java/banks/BankService.java
@@ -1,7 +1,5 @@
 package banks;
 
-import java.math.BigDecimal;
-
 public interface BankService {
-    public BigDecimal getRate(CurrencyName currencyName, int scale);
+    public double getRate(CurrencyName currencyName, int scale);
 }
